+++ index.html
@@ -6,13 +6,12 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/main.js"></script>
   <title>Кекстаграм</title>
-  <script src="js/function.js"></script>
 </head>
 
 <body>
   <script type="module" src="js/main.js"></script>
+  <script type="module" src="js/view.js"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
